+++ packages/twenty-front/src/modules/page-layout/widgets/components/WidgetSkeletonLoader.tsx
@@ -11,7 +11,7 @@ const StyledContainer = styled.div`
   width: 100%;
 `;
 
-export const ChartSkeletonLoader = () => {
+export const WidgetSkeletonLoader = () => {
   const theme = useTheme();
 
   return (
