+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidget.tsx
@@ -1,6 +1,6 @@
 import { useObjectMetadataItemById } from '@/object-metadata/hooks/useObjectMetadataItemById';
 import { PageLayoutWidgetNoDataDisplay } from '@/page-layout/widgets/components/PageLayoutWidgetNoDataDisplay';
-import { ChartSkeletonLoader } from '@/page-layout/widgets/graph/components/ChartSkeletonLoader';
+import { WidgetSkeletonLoader } from '@/page-layout/widgets/components/WidgetSkeletonLoader';
 import { GraphWidgetAggregateChartRenderer } from '@/page-layout/widgets/graph/graphWidgetAggregateChart/components/GraphWidgetAggregateChartRenderer';
 import { GraphWidgetBarChartRenderer } from '@/page-layout/widgets/graph/graphWidgetBarChart/components/GraphWidgetBarChartRenderer';
 import { GraphWidgetLineChartRenderer } from '@/page-layout/widgets/graph/graphWidgetLineChart/components/GraphWidgetLineChartRenderer';
@@ -54,7 +54,7 @@ export const GraphWidget = ({ objectMetadataId }: GraphWidgetProps) => {
       };
 
       return (
-        <Suspense fallback={<ChartSkeletonLoader />}>
+        <Suspense fallback={<WidgetSkeletonLoader />}>
           <GraphWidgetGaugeChart
             data={{
               value: gaugeData.value,
