+++ packages/twenty-server/src/engine/metadata-modules/front-component/front-component.service.ts
@@ -241,24 +241,67 @@ export class FrontComponentService {
     _id: string,
     _workspaceId: string,
   ): Promise<FrontComponentCodeDTO> {
+    // Worker code that communicates with the host via MessageChannel
+    // Uses the @remote-dom mutation protocol to send UI updates
     const sourceCode = `
-import { RemoteElement, RemoteRootElement } from '@remote-dom/core/elements';
+// Remote DOM mutation type constants (from @remote-dom/core)
+const MUTATION_TYPE_INSERT_CHILD = 0;
+const NODE_TYPE_ELEMENT = 1;
+const NODE_TYPE_TEXT = 3;
+const ROOT_ID = '~';
 
-customElements.define('remote-root', RemoteRootElement);
-customElements.define('remote-element', RemoteElement);
+let port = null;
+let idCounter = 0;
 
-const root = document.createElement('remote-root');
-document.body.appendChild(root);
+const generateId = () => String(++idCounter);
 
-const container = document.createElement('remote-element');
-container.setAttribute('type', 'div');
+const sendMutation = (records) => {
+  if (port) {
+    port.postMessage({ type: 'mutate', records });
+  }
+};
 
-const text = document.createElement('remote-element');
-text.setAttribute('type', 'span');
-text.textContent = 'Hello World from Custom Widget!';
+// Listen for the connect message from the host
+self.onmessage = (event) => {
+  if (event.data.type === 'connect') {
+    port = event.data.port;
+    port.start();
 
-container.appendChild(text);
-root.appendChild(container);
+    // Build and send the initial UI tree
+    const containerId = generateId();
+    const textId = generateId();
+
+    // Insert container element as child of root
+    sendMutation([
+      [
+        MUTATION_TYPE_INSERT_CHILD,
+        ROOT_ID,
+        {
+          id: containerId,
+          type: NODE_TYPE_ELEMENT,
+          element: 'remote-element',
+          properties: { type: 'div' },
+          children: [],
+        },
+        0,
+      ],
+    ]);
+
+    // Insert text node as child of container
+    sendMutation([
+      [
+        MUTATION_TYPE_INSERT_CHILD,
+        containerId,
+        {
+          id: textId,
+          type: NODE_TYPE_TEXT,
+          data: 'Hello World from Custom Widget!',
+        },
+        0,
+      ],
+    ]);
+  }
+};
 `;
 
     return { sourceCode };
